+++ visualization/app/codeCharta/services/loadInitialFile/loadInitialFile.service.ts
@@ -47,7 +47,7 @@ import { ExportCCFile } from "../../codeCharta.api.model"
 import { AppSettings, CcState, DynamicSettings, FileSettings, NameDataPair } from "../../codeCharta.model"
 import { FileState } from "../../model/files/files"
 import { getCCFiles } from "../../model/files/files.helper"
-import { MetricQueryParemter } from "../../state/effects/saveMetricsInQueryParameters/saveMetricsInQueryParameters.effect"
+import { MetricQueryParemter } from "../../state/effects/updateQueryParameters/updateQueryParameters.effect"
 import { metricDataSelector } from "../../state/selectors/accumulatedData/metricData/metricData.selector"
 import { setIsLoadingFile } from "../../state/store/appSettings/isLoadingFile/isLoadingFile.actions"
 import { setIsLoadingMap } from "../../state/store/appSettings/isLoadingMap/isLoadingMap.actions"
@@ -57,6 +57,7 @@ import { buildHtmlMessage } from "../../util/loadFilesValidationToErrorDialog"
 import { getNameDataPair } from "../loadFile/fileParser"
 import { LoadFileService, NO_FILES_LOADED_ERROR_MESSAGE } from "../loadFile/loadFile.service"
 import { UrlExtractor } from "./urlExtractor"
+import { setCurrentFilesAreSampleFiles } from "../../state/store/appStatus/currentFilesAreSampleFiles/currentFilesAreSampleFiles.actions"
 
 export const sampleFile1 = { fileName: "sample1.cc.json", fileSize: 3 * 1024, content: sample1 as ExportCCFile }
 export const sampleFile2 = { fileName: "sample2.cc.json", fileSize: 2 * 1024, content: sample2 as ExportCCFile }
@@ -106,6 +107,7 @@ export class LoadInitialFileService {
             await this.handleErrorLoadFilesFromQueryParams(error as Error)
         } finally {
             this.setMetricsFromUrl()
+            this.setCurrentFilesAreSampleFilesFromUrl()
         }
     }
 
@@ -417,6 +419,7 @@ export class LoadInitialFileService {
         } catch {
             this.loadFileService.loadFiles([sampleFile1, sampleFile2])
         }
+        this.store.dispatch(setCurrentFilesAreSampleFiles({ value: true }))
     }
 
     private showErrorDialog(title: string, message: string) {
@@ -473,6 +476,13 @@ export class LoadInitialFileService {
 
         if (renderState === "Delta" && files.length >= 2) {
             this.store.dispatch(setDelta({ referenceFile: files[0], comparisonFile: files[1] }))
+        }
+    }
+
+    private setCurrentFilesAreSampleFilesFromUrl() {
+        const currentFilesAreSampleFiles = this.urlUtils.getParameterByName(MetricQueryParemter.currentFilesAreSampleFiles)
+        if (currentFilesAreSampleFiles && currentFilesAreSampleFiles === "true") {
+            this.store.dispatch(setCurrentFilesAreSampleFiles({ value: true }))
         }
     }
 }