+++ visualization/app/codeCharta/util/dataMocks.ts
@@ -2122,6 +2122,7 @@ export const STATE: CcState = {
     },
     files: [],
     appStatus: {
+        currentFilesAreSampleFiles: false,
         hoveredNodeId: null,
         selectedBuildingId: null,
         rightClickedNodeData: null
@@ -2200,6 +2201,7 @@ export const DEFAULT_STATE: CcState = {
     },
     files: [],
     appStatus: {
+        currentFilesAreSampleFiles: false,
         hoveredNodeId: null,
         selectedBuildingId: null,
         rightClickedNodeData: null